Attempt fix to slowdown in full mode
It was taking a long time to open files because, I think, it was waiting for the query to finish. By clearing ('cancel' action) the query when opening I believe it will solve it. And also it's probably more useful to have fzf in a clean state for entrying new queries than the old way.